+++ eslint.config.js
@@ -29,9 +29,9 @@ export default [
       ...react.configs['jsx-runtime'].rules,
       ...reactHooks.configs.recommended.rules,
       'react/jsx-no-target-blank': 'off',
-      // "no-unused-vars": "off",
-      // 'react-hooks/exhaustive-deps': 'off',
-      // 'react/prop-types': 'off',
+      "no-unused-vars": "off",
+      'react-hooks/exhaustive-deps': 'off',
+      'react/prop-types': 'off',
       'react-refresh/only-export-components': [
         'warn',
         { allowConstantExport: true },
