+++ src/pages/order/OrderContext.jsx
@@ -1,0 +1,4 @@
+import { createContext } from "react";
+
+// 建立預約教練訂單的共用環境
+export const OrderContext = createContext({});
